Introducing DataFrame facade to limit amount of code duplication between Spark versions
