+++ datavec-spark/src/main/spark-2/org/datavec/spark/transform/BaseFlatMapFunctionAdaptee.java
@@ -5,6 +5,12 @@ import java.util.Iterator;
 import org.apache.spark.api.java.function.FlatMapFunction;
 import org.datavec.spark.functions.FlatMapFunctionAdapter;
 
+/**
+ * FlatMapFunction adapter to hide incompatibilities between Spark 1.x and Spark 2.x
+ *
+ * This class should be used instead of direct referral to FlatMapFunction
+ *
+ */
 public class BaseFlatMapFunctionAdaptee<K, V> implements FlatMapFunction<K,V> {
 
     protected final FlatMapFunctionAdapter<K, V> adapter;
