+++ datavec-spark/src/test/java/org/datavec/spark/transform/NormalizationTests.java
@@ -1,7 +1,6 @@
 package org.datavec.spark.transform;
 
 import org.apache.spark.api.java.JavaRDD;
-import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.datavec.api.transform.schema.Schema;
 import org.datavec.api.writable.DoubleWritable;
@@ -47,7 +46,7 @@ public class NormalizationTests extends BaseSparkTest {
 
         Schema schema = builder.build();
         JavaRDD<List<Writable>> rdd = sc.parallelize(data);
-        Dataset<Row> dataFrame = DataFrames.toDataFrame(schema,rdd);
+        DataRowsFacade dataFrame = DataFrames.toDataFrame(schema,rdd);
 
         //assert equivalent to the ndarray pre processing
         NormalizerStandardize standardScaler = new NormalizerStandardize();
@@ -58,7 +57,7 @@ public class NormalizationTests extends BaseSparkTest {
         zeroToOne.fit(new DataSet(arr.dup(),arr.dup()));
         INDArray zeroToOnes = arr.dup();
         zeroToOne.transform(new DataSet(zeroToOnes,zeroToOnes));
-        List<Row> rows = Normalization.stdDevMeanColumns(dataFrame,dataFrame.columns());
+        List<Row> rows = Normalization.stdDevMeanColumns(dataFrame,dataFrame.get().columns());
         INDArray assertion = DataFrames.toMatrix(rows);
         //compare standard deviation
         assertTrue(standardScaler.getStd().equalsWithEps(assertion.getRow(0),1e-1));
@@ -93,10 +92,10 @@ public class NormalizationTests extends BaseSparkTest {
         assertEquals(schema,DataFrames.fromStructType(DataFrames.fromSchema(schema)));
         assertEquals(rdd.collect(),DataFrames.toRecords(DataFrames.toDataFrame(schema,rdd)).getSecond().collect());
 
-        Dataset<Row> dataFrame = DataFrames.toDataFrame(schema,rdd);
-        dataFrame.show();
-        Normalization.zeromeanUnitVariance(dataFrame).show();
-        Normalization.normalize(dataFrame).show();
+        DataRowsFacade dataFrame = DataFrames.toDataFrame(schema,rdd);
+        dataFrame.get().show();
+        Normalization.zeromeanUnitVariance(dataFrame).get().show();
+        Normalization.normalize(dataFrame).get().show();
 
         //assert equivalent to the ndarray pre processing
         NormalizerStandardize standardScaler = new NormalizerStandardize();
